rename service_config -> service
